requirements.txt
Django==4.2.1
Pillow==9.5.0
Pygments==2.15.1
arrow==1.2.3
boto3==1.26.142
boto3-stubs==1.26.142
celery==5.2.7
celery-redbeat==2.1.0
dj-database-url==2.0.0
django-allauth==0.54.0
django-braces==1.15.0
django-cache-url==3.4.4
django-configurations==2.4.1
django-cors-headers==4.0.0
django-extensions==3.2.1
django-import-export==3.2.0
django-post-office==3.6.3
django-redis==5.2.0
django-ses==3.5.0
django-storages==1.13.2
djangoql==0.17.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
flake8==6.0.0
flower==1.2.0
gunicorn==20.1.0
honeybadger==0.15.2
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
sentry-sdk==1.24.0
whitenoise==6.4.0
django-smart-selects==1.6.0
